+++ src/map/entities/charentity.cpp
@@ -886,7 +886,7 @@ void CCharEntity::RemoveTrust(CTrustEntity* PTrust)
     {
         if (hasParty())
         {
-            getParty().ipc().RemoveMember(PTrust->id);
+            getParty().ipc().removeMember(PTrust->id);
         }
 
         if (PTrust->animation == ANIMATION_DESPAWN)
@@ -904,7 +904,7 @@ void CCharEntity::ClearTrusts()
     {
         if (hasParty())
         {
-            getParty().ipc().RemoveMember(PTrust->id);
+            getParty().ipc().removeMember(PTrust->id);
         }
 
         PTrust->PAI->Despawn();
